refactor: 예약 api 폴더 manager과 guest로 분리
