--- frontend/src/api/map.ts
@@ -1,40 +1,0 @@
-import { AxiosResponse } from 'axios';
-import { QueryFunction, QueryKey } from 'react-query';
-import { QueryMapsSuccess, QueryMapSuccess } from 'types/response';
-import api from './api';
-
-export interface QueryGuestMapParams {
-  sharingMapId: string;
-}
-interface PostMapParams {
-  mapName: string;
-  mapDrawing: string;
-  mapImageSvg: string;
-}
-
-interface DeleteMapParams {
-  mapId: number;
-}
-
-export const queryManagerMaps: QueryFunction<AxiosResponse<QueryMapsSuccess>> = () =>
-  api.get('/managers/maps');
-
-export const queryGuestMap: QueryFunction<
-  AxiosResponse<QueryMapSuccess>,
-  [QueryKey, QueryGuestMapParams]
-> = ({ queryKey }) => {
-  const [, data] = queryKey;
-  const { sharingMapId } = data;
-
-  return api.get(`/guests/maps?sharingMapId=${sharingMapId}`);
-};
-
-export const postMap = ({
-  mapName,
-  mapDrawing,
-  mapImageSvg,
-}: PostMapParams): Promise<AxiosResponse<never>> =>
-  api.post('/managers/maps', { mapName, mapDrawing, mapImageSvg });
-
-export const deleteMap = ({ mapId }: DeleteMapParams): Promise<AxiosResponse<never>> =>
-  api.delete(`/managers/maps/${mapId}`);
