+++ frontend/src/hooks/useGuestReservations.ts
@@ -1,12 +1,12 @@
 import { AxiosError, AxiosResponse } from 'axios';
 import { QueryKey, useQuery, UseQueryOptions, UseQueryResult } from 'react-query';
-import { queryGuestReservations, QuerySpaceReservationsParams } from 'api/reservation';
-import { ErrorResponse, QueryReservationsSuccess } from 'types/response';
+import { queryGuestReservations, QuerySpaceReservationsParams } from 'api/guestReservation';
+import { ErrorResponse, QueryGuestReservationsSuccess } from 'types/response';
 
-const useReservations = <TData = AxiosResponse<QueryReservationsSuccess>>(
+const useGuestReservations = <TData = AxiosResponse<QueryGuestReservationsSuccess>>(
   { mapId, spaceId, date }: QuerySpaceReservationsParams,
   options?: UseQueryOptions<
-    AxiosResponse<QueryReservationsSuccess>,
+    AxiosResponse<QueryGuestReservationsSuccess>,
     AxiosError<ErrorResponse>,
     TData,
     [QueryKey, QuerySpaceReservationsParams]
@@ -14,4 +14,4 @@ const useReservations = <TData = AxiosResponse<QueryReservationsSuccess>>(
 ): UseQueryResult<TData, AxiosError<ErrorResponse>> =>
   useQuery(['getReservations', { mapId, spaceId, date }], queryGuestReservations, options);
 
-export default useReservations;
+export default useGuestReservations;
