+++ frontend/src/pages/GuestReservationEdit/GuestReservationEdit.tsx
@@ -2,7 +2,7 @@ import { AxiosError } from 'axios';
 import { FormEventHandler } from 'react';
 import { useMutation } from 'react-query';
 import { useHistory, useLocation, useParams } from 'react-router-dom';
-import { putReservation } from 'api/reservation';
+import { putReservation } from 'api/guestReservation';
 import { ReactComponent as CalendarIcon } from 'assets/svg/calendar.svg';
 import Button from 'components/Button/Button';
 import Header from 'components/Header/Header';
@@ -12,8 +12,8 @@ import ReservationListItem from 'components/ReservationListItem/ReservationListI
 import MESSAGE from 'constants/message';
 import REGEXP from 'constants/regexp';
 import RESERVATION from 'constants/reservation';
+import useGuestReservations from 'hooks/useGuestReservations';
 import useInput from 'hooks/useInput';
-import useReservations from 'hooks/useReservations';
 import { GuestMapState } from 'pages/GuestMap/GuestMap';
 import { MapItem, Reservation, Space } from 'types/common';
 import { ErrorResponse } from 'types/response';
@@ -53,7 +53,7 @@ const GuestReservationEdit = (): JSX.Element => {
   const startDateTime = new Date(`${date}T${startTime}Z`);
   const endDateTime = new Date(`${date}T${endTime}Z`);
 
-  const getReservations = useReservations({ mapId, spaceId, date });
+  const getReservations = useGuestReservations({ mapId, spaceId, date });
   const reservations = getReservations.data?.data?.reservations ?? [];
 
   const editReservation = useMutation(putReservation, {
